chore: use deprecation dumpster for remaining deprecated decorator
